+++ core/exportutils.py
@@ -16,6 +16,7 @@ from .import_csv import (
 
 # The max lenght for an Excel tab name is 31. So truncate the name, if needed
 EXC_TAB_NAME_LEN = 31
+EXCEL_TYPE = "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
 
 
 def get_fk_value(obj, field, novalue="-"):
@@ -142,9 +143,6 @@ def generic_export_to_csv(queryset):
     return export_to_csv(queryset, generic_table_iterator)
 
 
-EXCEL_TYPE = "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
-
-
 def export_to_excel(queryset, f, title=""):
     if title == "":
         title = queryset.model._meta.verbose_name_plural.title()
@@ -164,62 +162,6 @@ def export_to_excel(queryset, f, title=""):
 
 def generic_export_to_excel(queryset):
     return export_to_excel(queryset, generic_table_iterator)
-
-
-def query_iterator(queryset, columns_dict):
-    month_list = ['Budget',
-                  'Apr',
-                  'May',
-                  'Jun',
-                  'Jul',
-                  'Aug',
-                  'Sep',
-                  'Oct',
-                  'Nov',
-                  'Dec',
-                  'Jan',
-                  'Feb',
-                  'Mar']
-    l = list(columns_dict.values())
-    l.extend(month_list)
-    yield l
-
-    for obj in queryset:
-        row = []
-        for field in columns_dict.keys():
-            val = obj[field]
-            if val is None:
-                val = ""
-            # row.append(str(val).encode("utf-8"))
-            row.append(val)
-        row.append(obj['Budget']/100)
-        row.append(obj['Apr']/100)
-        row.append(obj['May']/100)
-        row.append(obj['Jun']/100)
-        row.append(obj['Jul']/100)
-        row.append(obj['Aug']/100)
-        row.append(obj['Sep']/100)
-        row.append(obj['Oct']/100)
-        row.append(obj['Nov']/100)
-        row.append(obj['Dec']/100)
-        row.append(obj['Jan']/100)
-        row.append(obj['Feb']/100)
-        row.append(obj['Mar']/100)
-        yield row
-
-
-def export_query_to_excel(queryset, columns_dict, title):
-    resp = HttpResponse(content_type=EXCEL_TYPE)
-    filename = title + today_string() + " " + ".xlsx"
-    resp["Content-Disposition"] = "attachment; filename=" + filename
-    wb = openpyxl.Workbook()
-    ws = wb.get_active_sheet()
-    # Truncate the tab name to the maximum lenght permitted by Excel
-    ws.title = title[:EXC_TAB_NAME_LEN]
-    for row in query_iterator(queryset, columns_dict):
-        ws.append(row)
-    wb.save(resp)
-    return resp
 
 
 class export_csv_from_import:
