+++ forecast/utils/query_fields.py
@@ -6,6 +6,11 @@ BUDGET_TYPE_ORDER =  "financial_code__programme__budget_type_fk__budget_type_dis
 BUDGET_CATEGORY_ID = "financial_code__natural_account_code__expenditure_category__id"  # noqa
 BUDGET_CATEGORY_NAME = "financial_code__natural_account_code__expenditure_category__grouping_description"   # noqa
 
+# PAY, NON-PAY, CAPITAL, NON-CASH
+BUDGET_GROUPING = "financial_code__natural_account_code__expenditure_category__NAC_category__NAC_category_description"   # noqa
+BUDGET_NAC_CODE = "financial_code__natural_account_code__expenditure_category__linked_budget_code"  # noqa
+BUDGET_NAC_CODE_DESCRIPTION = "financial_code__natural_account_code__expenditure_category__linked_budget_code__natural_account_code_description"  # noqa
+
 # Admin, Capital or Programme
 FORECAST_EXPENDITURE_TYPE_NAME = "financial_code__forecast_expenditure_type__forecast_expenditure_type_name"  # noqa
 FORECAST_EXPENDITURE_TYPE_DESCRIPTION = "financial_code__forecast_expenditure_type__forecast_expenditure_type_description"  # noqa
@@ -25,6 +30,8 @@ GROUP_CODE = "financial_code__cost_centre__directorate__group__group_code"  # no
 
 NAC_CODE = "financial_code__natural_account_code__natural_account_code"
 NAC_NAME = "financial_code__natural_account_code__natural_account_code_description"  # noqa
+NAC_EXPENDITURE_TYPE = "financial_code__natural_account_code__economic_budget_code"  # noqa
+
 
 PROJECT_CODE = "financial_code__project_code__project_code"
 PROJECT_NAME = "financial_code__project_code__project_description"
@@ -35,10 +42,6 @@ ANALYSIS1_NAME = "financial_code__analysis1_code__analysis1_description"
 ANALYSIS2_CODE = "financial_code__analysis2_code__analysis2_code"
 ANALYSIS2_NAME = "financial_code__analysis2_code__analysis2_description"
 
-# PAY, NON-PAY, CAPITAL, NON-CASH
-BUDGET_GROUPING = "financial_code__natural_account_code__expenditure_category__NAC_category__NAC_category_description"   # noqa
-
-BUDGET_NAC_CODE = "financial_code__forecast_expenditure_type__forecast_expenditure_type_description"
 SHOW_DIT = 0
 SHOW_GROUP = 1
 SHOW_DIRECTORATE = 2
@@ -236,7 +239,7 @@ DEFAULT_PIVOT_COLUMNS = {
     PROJECT_NAME: "Project Description",
 }
 
-DOWNLOAD_COLUMNS = {
+VIEW_FORECAST_DOWNLOAD_COLUMNS = {
     GROUP_NAME: "Group name",
     GROUP_CODE: "Group code",
     DIRECTORATE_NAME : "Directorate Name" ,
@@ -244,13 +247,14 @@ DOWNLOAD_COLUMNS = {
     COST_CENTRE_NAME: "Cost Centre Description",
     COST_CENTRE_CODE: "Cost Centre Code",
     BUDGET_GROUPING: "Budget Grouping",
-    FORECAST_EXPENDITURE_TYPE_NAME: "FORECAST_EXPENDITURE_TYPE_NAME",
-    FORECAST_EXPENDITURE_TYPE_DESCRIPTION: "FORECAST_EXPENDITURE_TYPE_DESCRIPTION",
+    FORECAST_EXPENDITURE_TYPE_NAME: "Forecast Expenditure Type",
     BUDGET_TYPE: "Budget Type",
     BUDGET_CATEGORY_NAME: "Budget Category",
     BUDGET_NAC_CODE:"Budget/Forecast NAC",
+    BUDGET_NAC_CODE_DESCRIPTION: "Budget/Forecast NAC Description",
     NAC_CODE: "PO/Actual NAC",
     NAC_NAME: "Natural Account Code Description",
+    NAC_EXPENDITURE_TYPE: "NAC_EXPENDITURE_TYPE",
     PROGRAMME_CODE: "Programme Code",
     PROGRAMME_NAME: "Programme Description",
     ANALYSIS1_CODE: "Contract Code",
