+++ forecast/views/view_forecast/programme_details.py
@@ -72,7 +72,7 @@ class ForecastProgrammeDetailsMixin(MultiTableMixin):
             pivot_filter[filter_selectors[self.hierarchy_type]] = f"{filter_code}"
 
         columns = programme_details_hierarchy_columns[self.hierarchy_type]
-        programme_details_data = ForecastBudgetDataView.sub_total.subtotal_data(
+        programme_details_data = ForecastBudgetDataView.view_data.subtotal_data(
             programme_details_display_sub_total_column,
             programme_details_sub_total,
             columns.keys(),
